+++ 1D/mef_utilities/mef_process.hpp
@@ -1,4 +1,4 @@
-void create_local_K(Matrix* K, short element_id, Mesh* M){
+void create_local_K(Matrix* K, int element_id, Mesh* M){
     //(k/(x2-x1))*[1 -1; -1 1]
 
     K->set_size(2,2);
@@ -12,7 +12,7 @@ void create_local_K(Matrix* K, short element_id, Mesh* M){
     //cout << "\t\tLocal matrix created for Element " << element_id+1 << ": "; K->show(); cout << "\n";
 }
 
-void create_local_b(Vector* b, short element_id, Mesh* M){
+void create_local_b(Vector* b, int element_id, Mesh* M){
     //(Ql/2)[1;1]
     
     b->set_size(2);
@@ -26,7 +26,7 @@ void create_local_b(Vector* b, short element_id, Mesh* M){
     //cout << "\t\tLocal vector created for Element " << element_id+1 << ": "; b->show(); cout << "\n";
 }
 
-void create_local_systems(Matrix* Ks, Vector* bs, short num_elements, Mesh* M){
+void create_local_systems(Matrix* Ks, Vector* bs, int num_elements, Mesh* M){
     for(int e = 0; e < num_elements; e++){
         cout << "\tCreating local system for Element " << e+1 << "...\n\n";
         create_local_K(&Ks[e],e,M);
@@ -34,25 +34,25 @@ void create_local_systems(Matrix* Ks, Vector* bs, short num_elements, Mesh* M){
     }
 }
 
-void assembly_K(Matrix* K, Matrix* local_K, short index1, short index2){
+void assembly_K(Matrix* K, Matrix* local_K, int index1, int index2){
     K->add(local_K->get(0,0),index1,index1);    K->add(local_K->get(0,1),index1,index2);
     K->add(local_K->get(1,0),index2,index1);    K->add(local_K->get(1,1),index2,index2);
 }
 
-void assembly_b(Vector* b, Vector* local_b, short index1, short index2){
+void assembly_b(Vector* b, Vector* local_b, int index1, int index2){
     b->add(local_b->get(0),index1);
     b->add(local_b->get(1),index2);
 }
 
-void assembly(Matrix* K, Vector* b, Matrix* Ks, Vector* bs, short num_elements, Mesh* M){
+void assembly(Matrix* K, Vector* b, Matrix* Ks, Vector* bs, int num_elements, Mesh* M){
     K->init();
     b->init();
     //K->show(); b->show();
 
     for(int e = 0; e < num_elements; e++){
         cout << "\tAssembling for Element " << e+1 << "...\n\n";
-        short index1 = M->get_element(e)->get_node1()->get_ID() - 1;
-        short index2 = M->get_element(e)->get_node2()->get_ID() - 1;
+        int index1 = M->get_element(e)->get_node1()->get_ID() - 1;
+        int index2 = M->get_element(e)->get_node2()->get_ID() - 1;
 
         assembly_K(K, &Ks[e], index1, index2);
         assembly_b(b, &bs[e], index1, index2);
@@ -61,12 +61,12 @@ void assembly(Matrix* K, Vector* b, Matrix* Ks, Vector* bs, short num_elements, 
 }
 
 void apply_neumann_boundary_conditions(Vector* b, Mesh* M){
-    short num_conditions = M->get_quantity(NUM_NEUMANN);
+    int num_conditions = M->get_quantity(NUM_NEUMANN);
 
     for(int c = 0; c < num_conditions; c++){
         Condition* cond = M->get_neumann_condition(c);
         
-        short index = cond->get_node()->get_ID() - 1;
+        int index = cond->get_node()->get_ID() - 1;
         b->add(cond->get_value(), index);
     }
     //cout << "\t\t"; b->show(); cout << "\n";
@@ -78,7 +78,7 @@ void add_column_to_RHS(Matrix* K, Vector* b, int col, float T_bar){
 }
 
 void apply_dirichlet_boundary_conditions(Matrix* K, Vector* b, Mesh* M){
-    short num_conditions = M->get_quantity(NUM_DIRICHLET);
+    int num_conditions = M->get_quantity(NUM_DIRICHLET);
     int previous_removed = 0;
 
     for(int c = 0; c < num_conditions; c++){
