+++ 2D/geometry/node.hpp
@@ -1,20 +1,20 @@
 class Node {
     private:
-        short ID;
+        int ID;
         float x_coordinate;
         float y_coordinate;
 
     public:
-        Node(short identifier, float x_value, float y_value){
+        Node(int identifier, float x_value, float y_value){
             ID = identifier;
             x_coordinate = x_value;
             y_coordinate = y_value;
         }
 
-        void set_ID(short identifier){
+        void set_ID(int identifier){
             ID = identifier;
         }
-        short get_ID(){
+        int get_ID(){
             return ID;
         }
